+++ app/Http/Controllers/MenuController.php
@@ -9,8 +9,6 @@ use Illuminate\Http\Response;
 use Kami\Cocktail\Models\Menu;
 use OpenApi\Attributes as OAT;
 use Kami\Cocktail\OpenAPI as BAO;
-use Maatwebsite\Excel\Facades\Excel;
-use Kami\Cocktail\Exports\MenusExport;
 use Kami\Cocktail\Models\MenuCocktail;
 use Kami\Cocktail\Http\Requests\MenuRequest;
 use Kami\Cocktail\Http\Resources\MenuResource;