+++ app/OpenAPI/Schemas/CocktailPrice.php
@@ -19,10 +19,11 @@ class CocktailPrice
     public Price $total_price;
 
     /** @var array<mixed> */
-    #[OAT\Property(type: 'array', items: new OAT\Items(type: 'object', required: ['ingredient', 'price_per_amount', 'price_per_pour'], properties: [
+    #[OAT\Property(type: 'array', items: new OAT\Items(type: 'object', required: ['ingredient', 'price_per_unit', 'price_per_use', 'units'], properties: [
         new OAT\Property(type: IngredientBasic::class, property: 'ingredient'),
-        new OAT\Property(type: Price::class, property: 'price_per_amount', description: 'Price per 1 unit of ingredient amount'),
-        new OAT\Property(type: Price::class, property: 'price_per_pour', description: 'Price per cocktail ingredient part'),
+        new OAT\Property(type: 'string', property: 'units', description: 'Units used for price calculation'),
+        new OAT\Property(type: Price::class, property: 'price_per_unit', description: 'Price per 1 unit of ingredient amount'),
+        new OAT\Property(type: Price::class, property: 'price_per_use', description: 'Price per cocktail ingredient part'),
     ]))]
     public array $prices_per_ingredient;
 }