+++ resources/views/md_recipe_template.blade.php
@@ -10,9 +10,9 @@
 
 ## Ingredients
 @foreach ($cocktail->ingredients as $ci)
-- {{ $ci->formatter->printIngredient() }}{{ $ci->note ? ' - ' . $ci->note : '' }}
+- {{ (new \Kami\Cocktail\Models\CocktailIngredientFormatter($ci->amount, $ci->ingredient->name, $ci->optional))->format() }}{{ $ci->note ? ' - ' . $ci->note : '' }}
 @foreach ($ci->substitutes as $sub)
-    - or {{ $sub->formatter->printIngredient() }}
+    - or {{ (new \Kami\Cocktail\Models\CocktailIngredientFormatter($sub->amount, $sub->ingredient->name))->format() }}
 @endforeach
 @endforeach